Rename parent directory for action bar
